requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.30.0
pillow==10.2.0
openai==1.10.0
chromadb==0.4.22
sentence-transformers==2.3.1
python-dotenv==1.0.0
numpy==1.26.3
requests==2.31.0
